added redis service

+++ src/main/java/com/project/journel/entity/database/Category.java
@@ -1,13 +1,9 @@
 package com.project.journel.entity.database;
 
-import java.util.List;
-
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -27,7 +23,7 @@ public class Category {
   @Id
   @GeneratedValue
   @Column(name = "id")
-  private Integer id;
+  private Long id;
 
   @Column(name = "category_name", nullable = false)
   private String categoryName;
